app/requirements.txt
gradio>=4.0.0
requests
python-dotenv
numpy
espeak-ng
